pyproject.toml
[tool.poetry]
name = "sun-itup"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
openai = "^1.11.0"
llama-index = "^0.9.46"
python-dotenv = "^1.0.1"
pypdf = "^4.0.1"
qdrant-client = "^1.7.3"
llama-index-vector-stores-postgres = "^0.1.1"
llama-index-embeddings-huggingface = "^0.1.1"
llama-index-llms-llama-cpp = "^0.1.1"
psycopg2-binary = "^2.9.9"
pgvector = "^0.2.5"
asyncpg = "^0.29.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.26"}
greenlet = "^3.0.3"
psycopg2 = "^2.9.9"
llama-index-llms-openai = "^0.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[virtualenvs]
in-project = true